+++ lightly/api/__init__.py
@@ -3,6 +3,7 @@
 # Copyright (c) 2020. Lightly AG and its affiliates.
 # All Rights Reserved
 
+from lightly.api import routes
 from lightly.api.routes.pip import get_version              # noqa: F401
 from lightly.api.upload import upload_images_from_folder    # noqa: F401
 from lightly.api.upload import upload_embeddings_from_csv   # noqa: F401
