+++ lightly/api/routes/users/datasets/samples/service.py
@@ -3,8 +3,35 @@
 # Copyright (c) 2020. Lightly AG and its affiliates.
 # All Rights Reserved
 
-from . import _prefix
-import lightly.api.utils as utils
+from lightly.api.utils import getenv, get_request, post_request
+from typing import Union
+
+
+def _prefix(dataset_id: Union[str, None] = None,
+            sample_id: Union[str, None] = None,
+            *args, **kwargs):
+    """Returns the prefix for the samples routes.
+
+    Args:
+        dataset_id:
+            Identifier of the dataset.
+        sample_id:
+            Identifier of the sample.
+
+    """
+    server_location = getenv(
+        'LIGHTLY_SERVER_LOCATION',
+        'https://api.lightly.ai'
+    )
+    prefix = server_location + '/users/datasets'
+    if dataset_id is None:
+        prefix = prefix + '/samples'
+    else:
+        prefix = prefix + '/' + dataset_id + '/samples'
+    if sample_id is None:
+        return prefix
+    else:
+        return prefix + '/' + sample_id
 
 
 def get_presigned_upload_url(filename: str,
@@ -35,7 +62,7 @@ def get_presigned_upload_url(filename: str,
         'token': token
     }
 
-    response = utils.get_request(dst_url, params=payload)
+    response = get_request(dst_url, params=payload)
     signed_url = response.json()['signedWriteUrl']
     return signed_url
 
@@ -81,6 +108,6 @@ def post(filename: str,
     if thumbname is not None:
         payload['sample']['thumbName'] = thumbname
 
-    response = utils.post_request(dst_url, json=payload)
+    response = post_request(dst_url, json=payload)
     sample_id = response.json()['sampleId']
     return sample_id
