+++ lightly/api/routes/users/datasets/service.py
@@ -3,8 +3,27 @@
 # Copyright (c) 2020. Lightly AG and its affiliates.
 # All Rights Reserved
 
-from . import _prefix
-import lightly.api.utils as utils
+from typing import Union
+from lightly.api.utils import getenv, put_request
+
+
+def _prefix(dataset_id: Union[str, None] = None, *args, **kwargs):
+    """Returns the prefix for the datasets routes.
+
+    Args:
+        dataset_id:
+            The identifier of the dataset.
+
+    """
+    server_location = getenv(
+        'LIGHTLY_SERVER_LOCATION',
+        'https://api.lightly.ai'
+    )
+    prefix = server_location + '/users/datasets'
+    if dataset_id is None:
+        return prefix
+    else:
+        return prefix + '/' + dataset_id
 
 
 def put_image_type(dataset_id: str,
@@ -39,5 +58,5 @@ def put_image_type(dataset_id: str,
         }
     }
 
-    response = utils.put_request(dst_url, json=data, params=payload)
+    response = put_request(dst_url, json=data, params=payload)
     return response
